drop static smem analysis functionality and consolidate bank conflict logic
