+++ tests/compiler/diagnostics/test_bank_conflicts.py
@@ -1,377 +1,70 @@
-"""Tests for the smem bank-conflict simulator.
+"""Cross-validation: GPU bank-conflict trace vs the static oracle.
 
-Each fixture builds a minimal ``TileOp`` with a known ``Stage`` + body
-``Load`` and asserts the predicted lane→bank distribution.
-
-Expected ``max_way`` values are derived from first principles. To
-verify on real hardware, dump the equivalent kernel and run::
-
-    ncu --metrics smsp__sass_l1tex_data_bank_conflicts_pipe_lsu_mem_shared_op_ld.sum
-
-A 32-way conflict in this simulator predicts 31 bank-conflict events
-per warp-LDS instruction.
+For each ``(Stage, body Load)`` binding produced by the Tile-IR pipeline,
+asks ``lane_bank_distribution`` (the same oracle that ``009_permute_register_tile``
+and ``014_pad_smem`` score against) what each lane's smem address should be at
+``k_iter=0``, and checks ``simulate_graph`` records the same addresses.
+CUDA-only — skipped when no GPU is available.
 """
 
 from __future__ import annotations
 
-import pytest
-
-from deplodock.compiler.diagnostics.bank_conflicts import (
-    BANKS,
-    WARP_SIZE,
-    StageBinding,
-    simulate,
-    simulate_graph,
-    thread_axis_env,
-)
-from deplodock.compiler.graph import Graph
-from deplodock.compiler.ir.axis import BIND_THREAD, Axis, BoundAxis
-from deplodock.compiler.ir.expr import Literal, Var
-from deplodock.compiler.ir.stmt import Load, Loop, Tile
-from deplodock.compiler.ir.tile.ir import AffineAddressing, Stage, TileOp
-
-# ---------------------------------------------------------------------------
-# Fixture builder — a 1-Stage, 1-Load TileOp wrapped in a Tile with a
-# single thread axis (so threadIdx.x maps directly to that axis).
-# ---------------------------------------------------------------------------
+from tests.compiler.conftest import requires_cuda
 
 
-def _build_tileop(
-    *,
-    thread_axes: tuple[BoundAxis, ...],
-    stage_axes: tuple[Axis, ...],
-    pad: tuple[int, ...] = (),
-    load_index,
-    inner_loop_axis: Axis | None = None,
-    stage_cls=Stage,
-) -> TileOp:
-    """Build ``TileOp(Tile(Stage; [Loop?](Load)))``.
-
-    ``load_index`` is a tuple of ``Expr`` matching the staged buffer's
-    expected access (rank == ``len(stage_axes)``).
-    ``inner_loop_axis`` wraps the Load in a serial Loop — used when the
-    test wants to bind a k-iter; pass ``None`` for a Load directly in
-    the Tile body.
-    """
-    addressing = AffineAddressing(dims=tuple(range(len(stage_axes))))
-    stage = stage_cls(
-        name="smem",
-        buf="src",
-        origin=tuple(Literal(0) for _ in range(len(stage_axes))),
-        axes=stage_axes,
-        addressing=addressing,
-        pad=pad,
+@requires_cuda
+def test_oracle_matches_trace_on_small_matmul():
+    from deplodock.compiler.diagnostics.bank_conflicts import (
+        find_all_bindings,
+        lane_bank_distribution,
+        simulate_graph,
     )
-    load = Load(name="x", input="smem", index=tuple(load_index))
-    if inner_loop_axis is not None:
-        body = (stage, Loop(axis=inner_loop_axis, body=(load,)))
-    else:
-        body = (stage, load)
-    tile = Tile(axes=thread_axes, body=body)
-    return TileOp(body=(tile,), name="probe")
+    from deplodock.compiler.graph import Graph, Tensor
+    from deplodock.compiler.ir.base import InputOp
+    from deplodock.compiler.ir.frontend.ir import MatmulOp
+    from deplodock.compiler.pipeline import TILE_PASSES, run_pipeline
 
-
-def _binding_of(tile_op: TileOp) -> StageBinding:
-    from deplodock.compiler.diagnostics.bank_conflicts import find_stage_bindings
-
-    bindings = find_stage_bindings(tile_op)
-    assert len(bindings) == 1, f"expected exactly one binding, got {len(bindings)}"
-    return bindings[0]
-
-
-# ---------------------------------------------------------------------------
-# Direct unit tests for thread_axis_env (mirrors the runtime decode).
-# ---------------------------------------------------------------------------
-
-
-def test_thread_axis_env_single_axis():
-    ax = (Axis("a", 32),)
-    assert thread_axis_env(ax, 0) == {"a": 0}
-    assert thread_axis_env(ax, 31) == {"a": 31}
-
-
-def test_thread_axis_env_two_axes_outer_first():
-    """``Tile.thread_axes = (outer:8, inner:4)`` flattens via tid =
-    outer*4 + inner. Lane 0 → (0,0); lane 4 → (1,0); lane 7 → (1,3)."""
-    ax = (Axis("o", 8), Axis("i", 4))
-    assert thread_axis_env(ax, 0) == {"o": 0, "i": 0}
-    assert thread_axis_env(ax, 4) == {"o": 1, "i": 0}
-    assert thread_axis_env(ax, 7) == {"o": 1, "i": 3}
-    assert thread_axis_env(ax, 31) == {"o": 7, "i": 3}
-
-
-# ---------------------------------------------------------------------------
-# Bank-conflict fixtures
-# ---------------------------------------------------------------------------
-
-
-def test_row_strided_load_no_pad_is_32_way_conflict():
-    """All 32 lanes read different rows at fixed col=0 from a (32,32)
-    smem; row stride = 32 = #banks → every lane lands on bank 0."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        pad=(),
-        load_index=(Var("a"), Literal(0)),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 32
-    assert r.lane_banks == [0] * WARP_SIZE
-    assert r.counts[0] == 32
-
-
-def test_row_strided_load_pad_one_clears_conflict():
-    """Same access, pad=(0,1) makes row stride 33; bank = (lane*33)%32 =
-    lane%32, all distinct → no conflict."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        pad=(0, 1),
-        load_index=(Var("a"), Literal(0)),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 1
-    assert sorted(r.lane_banks) == list(range(32))
-
-
-def test_col_coalesced_load_no_conflict():
-    """All lanes read row 0 at column = lane → bank = lane → 1-way."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        pad=(),
-        load_index=(Literal(0), Var("a")),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 1
-
-
-def test_register_tile_F4_row_stride_16_8way_conflict():
-    """Pattern matmul-A loads with F=4: lane reads row=lane*4 in a
-    (128,16) smem at fixed col=0. row_stride=16 → bank=(lane*4*16)%32 =
-    (lane*64)%32 = 0 → 32-way."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 128), Axis("c", 16)),
-        pad=(),
-        load_index=(Var("a") * Literal(4), Literal(0)),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 32
-    assert r.counts[0] == 32
-
-
-def test_register_tile_F4_pad_one_partial_fix():
-    """Same F=4 pattern with pad=(0,1) → row stride 17.
-    addr = lane*4*17 = lane*68; bank = (lane*68)%32 = (lane*4)%32.
-    lane*4 mod 32 cycles through {0,4,8,...,28} (8 banks) with 4 lanes
-    each → max_way 4 (mild conflict — pad=1 is not enough for F=4)."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 128), Axis("c", 16)),
-        pad=(0, 1),
-        load_index=(Var("a") * Literal(4), Literal(0)),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 4
-    nonzero = [c for c in r.counts if c > 0]
-    assert len(nonzero) == 8
-    assert all(c == 4 for c in nonzero)
-
-
-def test_two_thread_axes_decode_into_index():
-    """``Tile.thread_axes=(o:8, i:4)`` flattens via tid=o*4+i.
-    Stage(8,4); Load `[o,i]` → lane reads `[lane//4, lane%4]`.
-    addr = (lane//4)*4 + (lane%4) = lane → bank = lane → 1-way."""
-    o = Axis("o", 8)
-    i = Axis("i", 4)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(o, BIND_THREAD), BoundAxis(i, BIND_THREAD)),
-        stage_axes=(Axis("r", 8), Axis("c", 4)),
-        pad=(),
-        load_index=(Var("o"), Var("i")),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    assert r.max_way == 1
-    assert sorted(r.lane_banks) == list(range(32))
-
-
-def test_k_iter_changes_bank_for_col_coalesced():
-    """Load `[0, a + k_iter]`: shifting k_iter shifts the bank by k_iter
-    but the per-lane spread is unchanged (still 1-way). Confirms k_iter
-    propagates into the eval env."""
-    a = Axis("a", 32)
-    k = Axis("k", 4)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        pad=(),
-        load_index=(Literal(0), Var("a") + Var("k")),
-        inner_loop_axis=k,
-    )
-    binding = _binding_of(tile_op)
-    r0 = simulate(binding, k_iter=0)
-    r2 = simulate(binding, k_iter=2)
-    assert r0 is not None and r2 is not None
-    assert r0.max_way == 1 and r2.max_way == 1
-    # Each lane shifts by 2 banks when k_iter goes 0→2.
-    assert r2.lane_banks == [(b + 2) % BANKS for b in r0.lane_banks]
-
-
-def test_buffered_stage_slot_dim_dropped():
-    """When the Load index has a leading slot dim that exceeds Stage.axes
-    rank, simulator drops it (slot is CTA-uniform, doesn't change banks).
-    Verifies the slot=0 vs slot=1 case yields identical bank spread."""
-
-    a = Axis("a", 32)
-    base_kwargs = dict(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        pad=(),
-    )
-    tile_op_slot0 = _build_tileop(**base_kwargs, load_index=(Literal(0), Var("a"), Literal(0)))
-    tile_op_slot1 = _build_tileop(**base_kwargs, load_index=(Literal(1), Var("a"), Literal(0)))
-    r0 = simulate(_binding_of(tile_op_slot0))
-    r1 = simulate(_binding_of(tile_op_slot1))
-    assert r0 is not None and r1 is not None
-    assert r0.lane_banks == r1.lane_banks
-    assert r0.max_way == 32  # row-strided → 32-way as in fixture 1
-
-
-def test_simulate_graph_walks_multiple_tileops():
-    """``simulate_graph`` should walk every TileOp in a Graph."""
     g = Graph()
-    a = Axis("a", 32)
-    for nid in ("k0", "k1"):
-        op = _build_tileop(
-            thread_axes=(BoundAxis(a, BIND_THREAD),),
-            stage_axes=(Axis("r", 32), Axis("c", 32)),
-            load_index=(Var("a"), Literal(0)),
-        )
-        op.name = nid
-        g.add_node(op, [], None, node_id=nid)
-    results = simulate_graph(g)
-    # Two TileOps × one (Stage, Load) each = 2 results.
-    assert len(results) == 2
-    assert all(r.max_way == 32 for r in results)
+    g.add_node(InputOp(), [], Tensor("a", (256, 64)), node_id="a")
+    g.add_node(InputOp(), [], Tensor("b", (64, 256)), node_id="b")
+    g.add_node(MatmulOp(), ["a", "b"], Tensor("c", (256, 256)), node_id="c")
+    g.inputs = ["a", "b"]
+    g.outputs = ["c"]
 
+    tile_graph = run_pipeline(g, TILE_PASSES)
 
-def test_stage_filter_restricts_results():
-    """``stage_filter`` excludes non-matching stage names."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        load_index=(Var("a"), Literal(0)),
+    oracle_by_key: dict[tuple[str, str, str], tuple[list[int], list[int]]] = {}
+    for binding in find_all_bindings(tile_graph):
+        stage, load, tile = binding.stage, binding.load, binding.tile
+        if not stage.axes or len(load.index) < len(stage.axes):
+            continue
+        cache_idx = tuple(load.index[-len(stage.axes) :])
+        extra_env: dict[str, int] = {ax.name: 0 for ax in tile.block_axes}
+        for ax in binding.enclosing_loop_axes:
+            extra_env.setdefault(ax.name, 0)
+        dist = lane_bank_distribution(cache_idx, stage.alloc_extents, tile.thread_axes, extra_env=extra_env)
+        if dist is None:
+            continue
+        oracle_by_key[(binding.tile_op_name, stage.name, load.name)] = (dist.lane_addrs, dist.lane_banks)
+
+    trace_results = simulate_graph(tile_graph)
+    assert trace_results, "simulate_graph returned no results"
+    assert oracle_by_key, "oracle produced no predictions"
+
+    trace_by_key = {(r.tile_op_name, r.stage_name, r.load_name): r for r in trace_results}
+    common = set(oracle_by_key) & set(trace_by_key)
+    assert common, (
+        f"no overlapping (kernel, stage, load) keys between oracle and trace\n"
+        f"  oracle: {sorted(oracle_by_key)}\n  trace:  {sorted(trace_by_key)}"
     )
-    g = Graph()
-    g.add_node(tile_op, [], None, node_id="k0")
-    assert simulate_graph(g, stage_filter={"smem"}) != []
-    assert simulate_graph(g, stage_filter={"other"}) == []
 
+    mismatches: list[str] = []
+    for key in sorted(common):
+        oracle_addrs, oracle_banks = oracle_by_key[key]
+        d = trace_by_key[key]
+        if oracle_addrs != d.lane_addrs:
+            mismatches.append(f"{key}: lane_addrs differ\n  oracle: {oracle_addrs}\n  trace:  {d.lane_addrs}")
+        if oracle_banks != d.lane_banks:
+            mismatches.append(f"{key}: lane_banks differ\n  oracle: {oracle_banks}\n  trace:  {d.lane_banks}")
 
-def _build_multi_load_tileop(thread_axes, stage_axes, indices, stage_cls=Stage, pad=()):
-    addressing = AffineAddressing(dims=tuple(range(len(stage_axes))))
-    stage = stage_cls(
-        name="smem",
-        buf="src",
-        origin=tuple(Literal(0) for _ in range(len(stage_axes))),
-        axes=stage_axes,
-        addressing=addressing,
-        pad=pad,
-    )
-    loads = tuple(Load(name=f"x{i}", input="smem", index=tuple(idx)) for i, idx in enumerate(indices))
-    tile = Tile(axes=thread_axes, body=(stage, *loads))
-    return TileOp(body=(tile,), name="multi")
-
-
-def test_lds128_chain_absorbs_4way_contiguous_columns():
-    """4 column-contiguous Loads in one TileOp form an LDS.128 chain;
-    each cycle's per-bank distinct count of 4 is absorbed → 0 events."""
-    a = Axis("a", 32)
-    indices = [(Literal(0), Var("a") * Literal(4) + Literal(i)) for i in range(4)]
-    tile_op = _build_multi_load_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 32)),
-        indices=indices,
-    )
-    g = Graph()
-    g.add_node(tile_op, [], None, node_id="k")
-    results = simulate_graph(g)
-    assert len(results) == 4
-    assert all(r.vec_group_size == 4 for r in results), [r.vec_group_size for r in results]
-    # Per cycle: 32 lanes × address (a*4+i) → 8 distinct banks, 4 lanes
-    # per bank → 4 distinct addrs each (no broadcasts here since lane*4
-    # mod 32 is bijective on lanes 0..7 then repeats with different addr).
-    # max(0, 4 - vec_width=4) = 0 absorbed.
-    assert all(r.lds128_events == 0 for r in results), [r.lds128_events for r in results]
-
-
-def test_lds128_chain_does_not_absorb_8way_unaligned():
-    """Same row but lane stride 8 — each cycle's per-bank distinct count
-    is 8 > vec_width=4 → 4 events absorbed, 4 events remain per bank.
-    Unaligned with the LDS.128 width, so absorption is partial."""
-    a = Axis("a", 32)
-    indices = [(Literal(0), Var("a") * Literal(8) + Literal(i)) for i in range(4)]
-    tile_op = _build_multi_load_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 32), Axis("c", 64)),
-        indices=indices,
-    )
-    g = Graph()
-    g.add_node(tile_op, [], None, node_id="k")
-    results = simulate_graph(g)
-    assert len(results) == 4
-    assert all(r.vec_group_size == 4 for r in results)
-    # Per Load: 32 lanes × (a*8+i) hit 4 banks (since 8 mod 32 cycle
-    # length 4) with 8 distinct addrs each → 8-way per bank.
-    # absorbed = max(0, 8 - 4) = 4 events per bank × 4 active banks = 16.
-    assert all(r.lds128_events == 16 for r in results), [r.lds128_events for r in results]
-
-
-def test_lds128_non_vectorizable_row_strided_unchanged():
-    """Loads at the same column but consecutive rows — NOT vectorizable
-    (rows aren't contiguous). vec_group_size stays 1; lds128_events ==
-    conflict_events."""
-    a = Axis("a", 32)
-    # ``[a*4+i, 0]`` — 4 different rows, same col 0. Adjacent Loads
-    # differ by `+row_stride` per lane, not +1 → not vectorizable.
-    indices = [(Var("a") * Literal(4) + Literal(i), Literal(0)) for i in range(4)]
-    tile_op = _build_multi_load_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", 128), Axis("c", 16)),
-        indices=indices,
-    )
-    g = Graph()
-    g.add_node(tile_op, [], None, node_id="k")
-    results = simulate_graph(g)
-    assert len(results) == 4
-    assert all(r.vec_group_size == 1 for r in results)
-    assert all(r.lds128_events == r.conflict_events for r in results)
-
-
-@pytest.mark.parametrize("rows", [16, 32, 64, 128])
-def test_avg_lane_per_bank_consistent(rows: int):
-    """``avg_way`` equals 32/distinct_banks. Sanity: row-strided no-pad
-    always lands on 1 bank → avg=32."""
-    a = Axis("a", 32)
-    tile_op = _build_tileop(
-        thread_axes=(BoundAxis(a, BIND_THREAD),),
-        stage_axes=(Axis("r", rows), Axis("c", 32)),
-        load_index=(Var("a"), Literal(0)),
-    )
-    r = simulate(_binding_of(tile_op))
-    assert r is not None
-    nonzero = [c for c in r.counts if c > 0]
-    assert r.avg_way == pytest.approx(WARP_SIZE / len(nonzero))
+    assert not mismatches, "oracle/trace disagreement:\n" + "\n".join(mismatches)
